cleaned up src/models/dggrid.rs
