+++ src/models/mod.rs
@@ -8,4 +8,4 @@
 // except according to those terms.
 
 pub mod common;
-pub mod dggrid;
+//pub mod dggrid;
